requirements.txt
aiogram>=3.3.0
python-dotenv>=1.0.0
python-digitalocean>=1.17.0
paramiko>=3.4.0 
